refactor: 폴더 구조 개선 및 MyPage를 MyAccount로 변경
